Fixed Caffe feature encoding to work (bug with ordering of net input images) * Fix also applies to loading of image mean (previously incorrect) * Added debug image output to CaffeEncoder * Added debug code to output features to MAT file (with optional MATIO   dependency when this mode is enabled) * Also added converstion between BGR->RGB option for backwards   compatibility withvgg-caffe * Fixed unsigned comparison warnings * Moved all data to test_data/ subdirectory and made paths in binaries relative * TODO: fix feats/ensureOutputCorrect test, but output is now working   corectly

+++ src/server/util/io.cc
@@ -33,7 +33,7 @@ namespace cpuvisor {
     CHECK(feats.isContinuous());
 
     const float* feats_data = (float*)feats.data;
-    for (size_t i = 0; i < feats.rows*feats.cols; ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(feats.rows*feats.cols); ++i) {
       feats_proto.add_data(feats_data[i]);
     }
     for (size_t i = 0; i < paths.size(); ++i) {
